requirements.txt
behave-django==1.1.0
dj-database-url==0.5.0
Django==2.0
django-i18nfield==1.4.0
django-polymorphic==2.0.3
django-rest-polymorphic==0.1.8
djangorestframework==3.8.2
gunicorn==19.9.0
pip==18.1
psycopg2==2.7.5
pytz==2018.5
setuptools==40.5.0
wheel==0.32.2
whitenoise==4.0